fix: correctly get output spatial reference for POST requests
Includes a refactor to move the helper function out into the utils module.
